requirements.txt
asgiref==3.8.1
Django==5.1.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
pillow==10.4.0
psycopg2==2.9.9
sqlparse==0.5.1
tzdata==2024.1

djangorestframework~=3.15.2
drf-yasg~=1.21.7
cloudinary~=1.41.0